+++ propagation/http_trace_context_propagator.go
@@ -18,20 +18,24 @@ import (
 	"context"
 	"encoding/hex"
 	"fmt"
+	"net/url"
 	"regexp"
 	"strconv"
 	"strings"
 
-	"go.opentelemetry.io/api/trace"
+	"go.opentelemetry.io/api/key"
 
 	"go.opentelemetry.io/api/core"
+	dctx "go.opentelemetry.io/api/distributedcontext"
 	apipropagation "go.opentelemetry.io/api/propagation"
+	"go.opentelemetry.io/api/trace"
 )
 
 const (
-	supportedVersion  = 0
-	maxVersion        = 254
-	TraceparentHeader = "Traceparent"
+	supportedVersion         = 0
+	maxVersion               = 254
+	TraceparentHeader        = "Traceparent"
+	CorrelationContextHeader = "Correlation-Context"
 )
 
 // HTTPTraceContextPropagator propagates SpanContext in W3C TraceContext format.
@@ -40,7 +44,7 @@ type HTTPTraceContextPropagator struct{}
 var _ apipropagation.TextFormatPropagator = HTTPTraceContextPropagator{}
 var traceCtxRegExp = regexp.MustCompile("^[0-9a-f]{2}-[a-f0-9]{32}-[a-f0-9]{16}-[a-f0-9]{2}-?")
 
-func (hp HTTPTraceContextPropagator) Inject(ctx context.Context, supplier apipropagation.Supplier) {
+func (hp HTTPTraceContextPropagator) Inject(ctx context.Context, correlationCtx dctx.Map, supplier apipropagation.Supplier) {
 	sc := trace.CurrentSpan(ctx).SpanContext()
 	if sc.IsValid() {
 		h := fmt.Sprintf("%.2x-%.16x%.16x-%.16x-%.2x",
@@ -51,9 +55,34 @@ func (hp HTTPTraceContextPropagator) Inject(ctx context.Context, supplier apipro
 			sc.TraceFlags&core.TraceFlagsSampled)
 		supplier.Set(TraceparentHeader, h)
 	}
+
+	firstIter := true
+	var headerValueBuilder strings.Builder
+	correlationCtx.Foreach(func(kv core.KeyValue) bool {
+		if !firstIter {
+			headerValueBuilder.WriteRune(',')
+		}
+		firstIter = false
+		headerValueBuilder.WriteString(url.QueryEscape(strings.TrimSpace(kv.Key.Name)))
+		headerValueBuilder.WriteRune('=')
+		headerValueBuilder.WriteString(url.QueryEscape(strings.TrimSpace(kv.Value.Emit())))
+		return true
+	})
+	if headerValueBuilder.Len() > 0 {
+		headerString := headerValueBuilder.String()
+		supplier.Set(CorrelationContextHeader, headerString)
+	}
 }
 
-func (hp HTTPTraceContextPropagator) Extract(ctx context.Context, supplier apipropagation.Supplier) core.SpanContext {
+func (hp HTTPTraceContextPropagator) Extract(
+	ctx context.Context, supplier apipropagation.Supplier,
+) (core.SpanContext, dctx.Map) {
+	return hp.extractSpanContext(ctx, supplier), hp.extractCorrelationCtx(ctx, supplier)
+}
+
+func (hp HTTPTraceContextPropagator) extractSpanContext(
+	ctx context.Context, supplier apipropagation.Supplier,
+) core.SpanContext {
 	h := supplier.Get(TraceparentHeader)
 	if h == "" {
 		return core.EmptySpanContext()
@@ -128,6 +157,50 @@ func (hp HTTPTraceContextPropagator) Extract(ctx context.Context, supplier apipr
 	return sc
 }
 
+func (hp HTTPTraceContextPropagator) extractCorrelationCtx(ctx context.Context, supplier apipropagation.Supplier) dctx.Map {
+	correlationContext := supplier.Get(CorrelationContextHeader)
+	if correlationContext == "" {
+		return dctx.NewEmptyMap()
+	}
+
+	contextValues := strings.Split(correlationContext, ",")
+	keyValues := make([]core.KeyValue, 0, len(contextValues))
+	for _, contextValue := range contextValues {
+		valueAndProps := strings.Split(contextValue, ";")
+		if len(valueAndProps) < 1 {
+			continue
+		}
+		nameValue := strings.Split(valueAndProps[0], "=")
+		if len(nameValue) < 2 {
+			continue
+		}
+		name, err := url.QueryUnescape(nameValue[0])
+		if err != nil {
+			continue
+		}
+		trimmedName := strings.TrimSpace(name)
+		value, err := url.QueryUnescape(nameValue[1])
+		if err != nil {
+			continue
+		}
+		trimmedValue := strings.TrimSpace(value)
+
+		// TODO (skaris): properties defiend https://w3c.github.io/correlation-context/, are currently
+		// just put as part of the value.
+		var trimmedValueWithProps strings.Builder
+		trimmedValueWithProps.WriteString(trimmedValue)
+		for _, prop := range valueAndProps[1:] {
+			trimmedValueWithProps.WriteRune(';')
+			trimmedValueWithProps.WriteString(prop)
+		}
+
+		keyValues = append(keyValues, key.New(trimmedName).String(trimmedValueWithProps.String()))
+	}
+	return dctx.NewMap(dctx.MapUpdate{
+		MultiKV: keyValues,
+	})
+}
+
 func (hp HTTPTraceContextPropagator) GetAllKeys() []string {
-	return []string{TraceparentHeader}
+	return []string{TraceparentHeader, CorrelationContextHeader}
 }
